refactor (kitchen pos) : “MenuGroupDao 인터페이스 추출 및 기존 MenuGroupDao 클래스 이름을 JdbcTemplateMenuGroupDao로 변경
